requirements.txt
flask
flask-restx
python-dotenv
google-genai
numpy
chromadb
requests
PyMuPDF